// src/shared/model/types/searchMulti200ResultsItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * tmdb-api
 * OpenAPI spec version: 3
 */

export type SearchMulti200ResultsItem = {
  adult?: boolean;
  backdrop_path?: string;
  id?: number;
  title?: string;
  original_language?: string;
  original_title?: string;
  overview?: string;
  poster_path?: string;
  media_type?: string;
  genre_ids?: number[];
  popularity?: number;
  release_date?: string;
  video?: boolean;
  vote_average?: number;
  vote_count?: number;
  name?: string;
  original_name?: string;
};