// src/shared/model/types/discoverMovie200ResultsItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * tmdb-api
 * OpenAPI spec version: 3
 */

export type DiscoverMovie200ResultsItem = {
  adult?: boolean;
  backdrop_path?: string;
  genre_ids?: number[];
  id?: number;
  original_language?: string;
  original_title?: string;
  overview?: string;
  popularity?: number;
  poster_path?: string;
  release_date?: string;
  title?: string;
  video?: boolean;
  vote_average?: number;
  vote_count?: number;
};