// src/shared/model/types/discoverTv200ResultsItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * tmdb-api
 * OpenAPI spec version: 3
 */

export type DiscoverTv200ResultsItem = {
  backdrop_path?: string;
  first_air_date?: string;
  genre_ids?: number[];
  id?: number;
  name?: string;
  origin_country?: string[];
  original_language?: string;
  original_name?: string;
  overview?: string;
  popularity?: number;
  poster_path?: string;
  vote_average?: number;
  vote_count?: number;
};