// src/shared/model/types/searchPerson200ResultsItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * tmdb-api
 * OpenAPI spec version: 3
 */
import type { SearchPerson200ResultsItemKnownForItem } from './searchPerson200ResultsItemKnownForItem';

export type SearchPerson200ResultsItem = {
  adult?: boolean;
  gender?: number;
  id?: number;
  known_for_department?: string;
  name?: string;
  original_name?: string;
  popularity?: number;
  profile_path?: string;
  known_for?: SearchPerson200ResultsItemKnownForItem[];
};